+++ CHANGELOG.md
@@ -12,6 +12,11 @@ XCode Templates
 
 ### 3.8.0(30)
 
+Views and ViewControllers
+
+- `SecurableTextField` is added
+- `PlainDisclosureGroup` is added
+
 Global Functions
 
 - `KeyPath` equality and comparison functions are extended to support `10` members